pyproject.toml
[project]
name = "tinkoff-bot"
version = "0.1.0"
description = "Проект для торговли и отслеживаия инструментов на бирже"
authors = [
    { name = "Оплеснин Иван", email = "aplesnin2010@yandex.ru" }
]
readme = "README.md"
requires-python = ">=3.13,<4.0.0"
dependencies = [
    "flake8 (>=7.3.0,<8.0.0)",
    "tinkoff-investments (>=0.2.0b116,<0.3.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "aiogram (>=3.22.0,<4.0.0)",
    "hvac (>=2.3.0,<3.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "flake8 (>=7.3.0,<8.0.0)"
]


[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
select = ["E", "F", "W"]
exclude = [".git", "__pycache__", "build", "dist", ".venv"]
per-file-ignores = [
  "__init__.py:F401",
]